requirements.txt
Flask==2.0.2
opentelemetry-api==1.5.0
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-exporter-zipkin==1.6.2
opentelemetry-instrumentation==0.24b0
opentelemetry-instrumentation-flask==0.25b2
opentelemetry-sdk==1.5.0
requests==2.26.0
